# Block Editor: Flatten, shortcut align hook rendering if unaligned