@aduth 2019-12-06
@youknowriad I was having some difficulty with the performance test suite yesterday, but I was initially drawn to this code from a Chrome DevTools Performance recording, so there's some space for gains, likely in the 0-3ms range per keypress. I'll see if I can get the tests working today.