@aduth 2019-12-06
I'm still not entirely sure the performance tests are cooperating (and the results are a bit all over the place), but for what I do see, it doesn't seem like this is having too much of an impact for better or worse:

Before:

```
Average time to load: 4382ms
Average time to DOM content load: 4111ms
Average time to type character: 67.425ms
Slowest time to type character: 139ms
Fastest time to type character: 39ms
```

After:

```
Average time to load: 4190ms
Average time to DOM content load: 3915ms
Average time to type character: 67.39ms
Slowest time to type character: 136ms
Fastest time to type character: 38ms
```

Conceptually, I can't imagine how it could be anything other than an improvement, based on the proposed advantages here. At the very least it could be seen as as a useful refactoring to simplify the implementation of this hook, and toward consistency of hooks usage and component flattening.