@youknowriad 2019-12-06
@aduth if you want to push the performance tests further, you can try to disable the "Async Mode" which "virtually" improves performance and see if there's impact without it. (I don't think it's necessary though)